Fix the search dialog to not modify obsolete data.
If data in the page was changed from outside the search dialog, the
current search must be invalidated to avoid making edits (especially
via the text view's focus-out handler) based on data that no longer
exists. Invalidate the search if application emits a page-changed
signal, one that did not originate from the search dialog.

Fixes #572676.