Fix handling of Unicode BOMs.
Add 'utf_8_sig' (added with Python 2.5) to the list of supported
encodings with the name 'UTF-8-SIG'. Remove previous special-case
handling of UTF-8 BOMs in SubtitleFile and handle UTF-8 BOMs normally
via encodings.

Add function 'detect_bom' to gaupol.encodings and use that to check
files for Unicode BOMs. Do this check each time when opening a file
before trying the defined encodings to avoid getting a FormatError if
an encoding decodes the file into garbage without raising a
UnicodeDecodeError. If a BOM is found, read the file with the
correspoding encoding. This means that Unicode files with a BOM should
be always opened correctly regardless of the locale encoding or the
defined fallback encodings.

In SubtitleFile._read_lines and SubtitleFile.write handle the various
BOMs as follows.

 * UTF-8-SIG: If a UTF-8 BOM is found (i.e. UTF-8-SIG file opened as
   UTF-8), reread the file as UTF-8-SIG. Once UTF-8-SIG is used,
   Python automatically strips the BOM when reading and adds it when
   writing.

 * UTF-16: When using UTF-16, Python automatically strips the BOM
   (whether it is BE or LE) when reading and adds the plaform's native
   byte order BOM when writing.

 * UTF-16-BE and UTF-16-LE: When using these Python does no automatic
   stripping or adding of BOMs. Thus, when reading, set SubtitleFile
   attribute 'has_bom_utf_16' and when writing, check that attribute
   and add the corresponding UTF-16 BOM if one was read.

Because the 'detect_bom' function is always used before other
encodings, the possibly destructive, BOM-changing encoding UTF-16
should never end up used when opening files.