Change the order in which patterns are run.
Previously patterns have been run in the order they are defined in the
pattern file with general patterns run first, then script patterns,
then language patterns and finally country patterns. Change this so
that patterns with the same name are always run one after the other in
the position of the earliest one.

For example, consider the following patterns defined in the pattern
files in the following order.

    Latn.common-error:    A, B, C
    Latn-fi.common-error: B, D, E

Previously these were run in order A, B, C, B, D, E. With this change
they will now be run in order A, B, B, C, D, E or if the policy of the
B pattern in Latn-fi has been set to 'Replace', the A, B, C, D, E.
This should make it more convenient to replace or extend patterns
above in the hierarchy.

Adapt Latn-fr.common-error.in to the above order and add
Latn-fi.common-error.in to use the new order.