Restore zebra-stripes.
Since some GTK+ themes have chosen to ignore
Gtk.TreeViewColumn.set_rules_hint, we need to, unfortunately, be more
fierce settings those colors. For some users this is a big deal to
clarify multiline text. For now, zebra is the default, but that can be
changed later and perhaps even added to the preferences dialog.