Poll even when paused.
Activating and deactivating update handlers when pausing or restarting
playback makes many common operations such as seeking and segment
playback way too slow. And to some extent we need updates even when not
playing.

Polling is still probably insignificant compared to the CPU-cost of
video decoding and rendering, but it does feel a waste to poll when not
needed. Let's wait for complaints about this being horribly slow before
trying to optimize it further.