Add a framerate selector to save dialogs.
This is shown only when converting from a time-based file format to a
frame-based or vice versa and should be a helpful reminder that the
framerate needs to be set when doing such a conversion. Since the
removal of the video toolbar the only remaining framerate selector in
the main window is in the "View" menu, which is not obvious to affect
files written to disk.