Fix em width to be at a more correct level
To calculate the em width, let's just use an average width of 0.55 em
per a-z character since that we can understand and verify to work
somewhat correcly as opposed values from Pango. The font with which the
calculations are done remains whatever GTK+ considers to be the default
to be.