Drop the bookmarks and side-pane extensions
The codebases are way too complicated, which means that it would be too
much work for little benefit to port to Gio.Action/Gio.Menu etc. and
they have a history of being broken often due not being used much and
being hard to test.

These were mainly written to test and demonstrate the extension system
by doing lots of cool and thus pervasive things. The extension system
hasn't really caught on and the migrations to Python 3, PyGObject and
now Gio prove that it perhaps never made sense to expose internals for
public consumption (see a thread at desktop-devel-list [1] for a good
recap of the extension API problem in general).

[1] https://mail.gnome.org/archives/desktop-devel-list/2016-May/msg00017.html