Update video.py
The method player.get_audio_languages() can return None, but that case wasn't being dealt with. I got the error while trying to load a video in my system (archlinux) using gaupol-0.92-1, and it was getting in the way so I couldn't load any video. You may want to deal with this in a fancier manner but for the time being this is an easy fix that worked for me.