Fix error trying to undo when holding down Ctrl+Z
This fix basically ignores undo requests when one is already in
progress, -- e.g. pressing Ctrl+Z quickly twice, when the first action
is slow to undo, would not undo a second time. The alternative would be
to queue the requests, but that could easily result in the queue growing
excessively long.

Closes #38