Use gtksink instead of autovideosink
Hopefully this makes the video player work on Wayland too (#60) without
needing weird special casing in the code using waylandsink. As a
downside gtksink is likely to use a lot more CPU than autovideosink.

This might allow the video player to work on Windows too as we no longer
need a window handle, but with pygi-aio-3.24 it seemed to result in a
crash with "No accelerated IMDCT transform found."

Requiring gtksink bumps the GStreamer dependency to 1.6.