Improve preview error dialogs
Check if player is found before running command and, if not, show a
simpler and clearer error dialog. Note that aeidon.ProcessError doesn't
seem to ever be fired here due to the process being run via a shell
(non-win32 systems, for reasons that I don't recall). In the preview
error dialog also point to the preferences dialog.

Part of #136